Refactor: gather declaration generating functions in one file
